backend/requirements.txt
fastapi
uvicorn
python-chess
sqlalchemy
pydantic
pytest
httpx
python-multipart